java.nio.file.Path converter. remove some reflection. make the jdk 1.6 build work.
